requirements.txt
python==3.11.1
streamlit==1.36.0
langchain==0.2.6
langchain_community==0.2.6
langchain_google_genai==1.0.7
faiss-cpu==1.8.0
nltk==3.8.1
protobuf==5.27.2
PyPDF2==3.0.1
python-dotenv==1.0.1
python_docx==1.1.2
google-generativeai==0.7.1
